passando msgs e contrutor
